+++ src/components/datagrid/data_grid.tsx
@@ -694,19 +694,10 @@ export const EuiDataGrid: FunctionComponent<EuiDataGridProps> = (props) => {
   }, [focusedCell, contentRef]);
 
   useEffect(() => {
-    rowHeightUtils.computeStylesForGridCell(
-      {
-        cellPadding: gridStyles.cellPadding,
-        fontSize: gridStyles.fontSize,
-      },
-      rowHeightsOptions?.lineHeight
-    );
-  }, [
-    gridStyles.cellPadding,
-    gridStyles.fontSize,
-    rowHeightsOptions?.lineHeight,
-    rowHeightUtils,
-  ]);
+    rowHeightUtils.cacheStyles({
+      cellPadding: gridStyles.cellPadding,
+    });
+  }, [gridStyles.cellPadding, rowHeightUtils]);
 
   const classes = classNames(
     'euiDataGrid',